refactor: reload_config の中身をクラス化
